Address PR review: share dictation download ring, sync a11y notification, HC track, docs
